requirements.txt
appdirs==1.4.4

asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.36
botocore==1.17.36
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.4
django-bootstrap-form==3.4
django-bootstrap4==2.2.0
django-cleanup==5.0.0
django-cors-middleware==1.5.0
django-debug-toolbar==2.2
django-email-verification==0.0.2
django-forms-bootstrap==3.1.0
django-formtools==2.2
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-otp==0.9.4
django-phonenumber-field==3.0.1
django-redis==4.12.1
django-robots==4.0
django-s3-folder-storage==0.1
django-select2==7.4.2
-e git://github.com/digi604/django-smart-selects.git@f5f7b826a543da2e6eeac37729edff9b7b8e8a64#egg=django_smart_selects
django-storages==1.9.1
django-two-factor-auth==1.12.1
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
jmespath==0.10.0
mailerlite-api-python==0.2.0
oauthlib==3.1.0
phonenumbers==8.12.7
Pillow==7.2.0
pipenv==2020.6.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pycryptodome==3.9.8
python-dateutil==2.8.1
pytz==2020.1
qrcode==6.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
SecretStorage==2.3.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
whitenoise==5.1.0
YubiOTP==0.2.2.post1
zipp==3.1.0
